add keyboard binding to popup
